fix filter dialog and order in filter result.

+++ src/com/boding/model/domestic/Flight.java
@@ -169,7 +169,7 @@ public class Flight {
 	}
 	
 	public String getFlightPrice(){
-		return String.valueOf(this.cabins.get(0).getAdultPrice());
+		return String.valueOf((int)this.cabins.get(0).getAdultPrice());
 	}
 	
 	public int getFlightPriceInt(){
